# docs/swagger.yaml
definitions:
  dto.Activity:
    properties:
      content:
        type: string
      end_time:
        type: string
      place:
        type: string
      start_time:
        description: Time：自1970年1月1日00:00:00 UTC以来经过的毫秒数
        type: string
      title:
        type: string
    required:
    - content
    - end_time
    - place
    - start_time
    - title
    type: object
  dto.Login:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.OrgInfo:
    properties:
      logo_url:
        type: string
      password:
        description: 长度大于8
        type: string
    type: object
  dto.Token:
    properties:
      token:
        type: string
    required:
    - token
    type: object
host: tudo-api-test.ncuos.com
info:
  contact:
    email: 316851756@qq.com
  description: |-
    给出了请求方法，点击Model可以查看请求体模型及备注。
    查看响应体需要打开浏览器开发者工具，在页面接口初Try it out，填写数据然后Execute。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger API
  version: "1.0"
paths:
  /activity:
    get:
      parameters:
      - description: 活动id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 获取活动详情
      tags:
      - 活动系统
  /auth/activity:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 活动id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 社团账号删除活动
      tags:
      - 活动系统
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: ' '
        in: body
        name: JSON
        required: true
        schema:
          $ref: '#/definitions/dto.Activity'
      produces:
      - application/json
      responses: {}
      summary: 社团账号发布活动
      tags:
      - 活动系统
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 活动id
        in: query
        name: id
        required: true
        type: string
      - description: ' '
        in: body
        name: JSON
        required: true
        schema:
          $ref: '#/definitions/dto.Activity'
      produces:
      - application/json
      responses: {}
      summary: 社团账号修改活动
      tags:
      - 活动系统
  /auth/organization:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 获取组织的负责人管理的组织名称
      tags:
      - 用户系统
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: ' '
        in: body
        name: JSON
        required: true
        schema:
          $ref: '#/definitions/dto.OrgInfo'
      produces:
      - application/json
      responses: {}
      summary: 社团账号激活（注册）、修改
      tags:
      - 用户系统
  /auth/picture:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: file字段放图片数据
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses: {}
      summary: 图片上传接口
      tags:
      - 其它
  /auth/token:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 检验token是否有效，若有效返回用户角色和新token
      tags:
      - 用户系统
  /during-activity:
    get:
      parameters:
      - description: 上一次调用本接口得到的活动列表的最后一个活动的开始时间戳，第一次调用用当前时间戳
        in: query
        name: pre
        required: true
        type: string
      - description: 当前时间戳
        in: query
        name: now
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 分页获取正在进行的活动列表
      tags:
      - 活动系统
  /ended-activity:
    get:
      parameters:
      - description: 上一次调用本接口得到的活动列表的最后一个活动的结束时间戳，第一次调用用当前时间戳
        in: query
        name: pre
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 分页获取历史活动列表
      tags:
      - 活动系统
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: JSON
        required: true
        schema:
          $ref: '#/definitions/dto.Login'
      produces:
      - application/json
      responses: {}
      summary: 云家园账号或社团账号登录
      tags:
      - 用户系统
  /login/ncuos-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: body
        name: JSON
        required: true
        schema:
          $ref: '#/definitions/dto.Token'
      produces:
      - application/json
      responses: {}
      summary: 云家园账号token登录
      tags:
      - 用户系统
  /not-start-activity:
    get:
      parameters:
      - description: 上一次调用本接口得到的活动列表的最后一个活动的开始时间戳，第一次调用用当前时间戳
        in: query
        name: pre
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 分页获取未开始的活动列表
      tags:
      - 活动系统
  /recommend-activity:
    get:
      parameters:
      - description: 上一次调用本接口得到的活动列表的最后一个活动的开始时间戳，第一次调用用当前时间戳
        in: query
        name: pre
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 分页获取推荐活动列表
      tags:
      - 活动系统
  /user-info:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 获取用户信息
      tags:
      - 用户系统
swagger: "2.0"